Refactor, move handlers from main to separate package.
