+++ authorization_server/handlers/handlers_test.go
@@ -1,4 +1,4 @@
-package main
+package handlers
 
 import (
 	"net/http"
@@ -43,11 +43,10 @@ func TestHealthCheckHandler(t *testing.T) {
 	}
 }
 
-func TestCookieGenerator(t *testing.T)  {
+func TestCookieGenerator(t *testing.T) {
 	twentyCookieCharacters := regexp.MustCompile("^[\\w+_]{20}$")
 	token := randomToken()
 	if !twentyCookieCharacters.MatchString(token) {
-		t.Error("Expected /^[\\w+-]{20}$/, got: "+token)
+		t.Error("Expected /^[\\w+-]{20}$/, got: " + token)
 	}
-	// fmt.Println("New token: ", token)
 }
