+++ authorization_server/types/types.go
@@ -2,11 +2,11 @@ package types
 
 // ответ сервера на всякие действия.
 type ServerResponse struct {
-	Status string `json:"status"`
+	Status  string `json:"status"`
 	Message string `json:"message"`
 }
 
 type NewUserRegistration struct {
-	Login string `json:"login"`
+	Login    string `json:"login"`
 	Password string `json:"password"`
 }
